Final Year Project
This project concerns about the growing risks to the life of people.
This app was made when we witnessed a lot of Women facing the threat of
rape , murders, molestations etc. Using this android app one can get the
fastest assistance from the nearby help centers as well as from his/her
family member just by pressing the power button a few times.
In order to use this app user must create an account as well as has to
input the contact number of those people whose assistance is required at
the time of emergency.